Removes PartiQLLexer and associated tests
